project/requirements.txt
alembic==1.4.2
blinker==1.4
click==7.1.1
dominate==2.5.1
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
pkg-resources==0.0.0
python-dateutil==2.8.1
python-decouple==3.3
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.16
visitor==0.1.3
Werkzeug==1.0.1
WTForm==1.0
WTForms==2.2.1
